# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Albelli API",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
...
        "/api/login": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Authentication using email and password",
                "operationId": "2c10d30e6b6731f53ec42e305183c141",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "password"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "alecuba16@gmail.com",
                                    "password": "alex"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Logged succesfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JsonResponse"
                                },
                                "example": {
                                    "success": true,
                                    "message": "User signed in",
                                    "data": {
                                        "token": "1|c9iSWWNP8iNfl1ZfM13cisTsM3oJE5oSGcgeA3Wf",
                                        "name": "alex"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Error validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JsonResponse"
                                },
                                "example": {
                                    "success": false,
                                    "message": "Unauthorised.",
                                    "data": {
                                        "error": "Unauthorised"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Registers a new user",
                "operationId": "3033a10d6c8ff93dfe3e20f50fe9822d",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "email"
                                    },
                                    "password": {
                                        "type": "password"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "alex",
                                    "email": "alecuba16@gmail.com",
                                    "password": "alex"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Registered successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JsonResponse"
                                },
                                "example": {
                                    "success": true,
                                    "message": "User created successfully.",
                                    "data": {
                                        "token": "1|c9iSWWNP8iNfl1ZfM13cisTsM3oJE5oSGcgeA3Wf",
                                        "name": "alex"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Error registering",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JsonResponse"
                                },
                                "example": {
                                    "success": false,
                                    "message": "Error validation",
                                    "data": {
                                        "email": [
                                            "The email has already been taken."
                                        ]
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/logout": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Logout the current user session",
                "operationId": "f6c4ffd7eb694be3829262c402db90f6",
                "responses": {
                    "200": {
                        "description": "Logout successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JsonResponse"
                                },
                                "example": {
                                    "success": true,
                                    "message": "Logout successfully"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unable to logout",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JsonResponse"
                                },
                                "example": {
                                    "success": false,
                                    "message": "User not authenticated."
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/checkLogin": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Check if the current token is valid",
                "operationId": "d8057d7f7b9b6f8f3590c1ef410af9bb",
                "responses": {
                    "200": {
                        "description": "Token check successfully"
                    },
                    "403": {
                        "description": "Token is no longer valid"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "JsonResponse": {
                "title": "Json Response Model",
                "description": "Json Response Model",
                "properties": {
                    "success": {
                        "description": "Returns true if the action was completed succesfully",
                        "type": "bolean",
                        "example": true
                    },
                    "message": {
                        "description": "Includes any aditional message that the endpoint may generate with informative proposes",
                        "type": "string",
                        "example": "Inserted ok. The query took 100ms"
                    },
                    "data": {
                        "description": "Optional data returned by the API server goes here or an error message if success was false"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "JsonResponse"
                }
            },
...
        },
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "scheme": "bearer",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "bearerFormat": "JWT",
                "in": "header"
            }
        }
    },
    "security": [
        []
    ]
}